% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediatailor_package.R
\docType{package}
\name{paws.mediatailor-package}
\alias{paws.mediatailor}
\alias{paws.mediatailor-package}
\title{paws.mediatailor: AWS MediaTailor}
\description{
Use the AWS Elemental MediaTailor SDK to configure
    scalable ad insertion for your live and VOD content. With AWS
    Elemental MediaTailor, you can serve targeted ads to viewers while
    maintaining broadcast quality in over-the-top (OTT) video
    applications. For information about using the service, including
    detailed information about the settings covered in this guide, see the
    AWS Elemental MediaTailor User Guide.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
