% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediatailor_operations.R
\name{put_playback_configuration}
\alias{put_playback_configuration}
\title{Adds a new configuration to AWS Elemental MediaTailor}
\usage{
put_playback_configuration(AdDecisionServerUrl = NULL,
  CdnConfiguration = NULL, Name = NULL, SlateAdUrl = NULL,
  VideoContentSourceUrl = NULL)
}
\arguments{
\item{AdDecisionServerUrl}{The URL for the ad decision server (ADS). This includes the specification of static parameters and placeholders for dynamic parameters. AWS Elemental MediaTailor substitutes player-specific and session-specific parameters as needed when calling the ADS. Alternately, for testing you can provide a static VAST URL. The maximum length is 25000 characters.}

\item{CdnConfiguration}{The configuration for using a content delivery network (CDN), like Amazon CloudFront, for content and ad segment management.}

\item{Name}{The identifier for the configuration.}

\item{SlateAdUrl}{The URL for a high-quality video asset to transcode and use to fill in time that's not used by ads. AWS Elemental MediaTailor shows the slate to fill in gaps in media content. Configuring the slate is optional for non-VPAID configurations. For VPAID, the slate is required because AWS Elemental MediaTailor provides it in the slots that are designated for dynamic ad content. The slate must be a high-quality asset that contains both audio and video.}

\item{VideoContentSourceUrl}{The URL prefix for the master playlist for the stream, minus the asset ID. The maximum length is 512 characters.}
}
\description{
Adds a new configuration to AWS Elemental MediaTailor.
}
\examples{

}
