% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediatailor_operations.R
\name{delete_playback_configuration}
\alias{delete_playback_configuration}
\title{Deletes the configuration for the specified name}
\usage{
delete_playback_configuration(Name)
}
\arguments{
\item{Name}{[required] The identifier for the configuration.}
}
\description{
Deletes the configuration for the specified name.
}
\section{Accepted Parameters}{
\preformatted{delete_playback_configuration(
  Name = "string"
)
}
}

