% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.mediatailor_operations.R
\name{get_playback_configuration}
\alias{get_playback_configuration}
\title{Returns the configuration for the specified name}
\usage{
get_playback_configuration(
  Name = "string"
)
}
\arguments{
\item{Name}{[required] The identifier for the configuration.}
}
\description{
Returns the configuration for the specified name.
}
\examples{

}
